+++ lib/absinthe/phase/schema/type_imports.ex
@@ -35,7 +35,7 @@ defmodule Absinthe.Phase.Schema.TypeImports do
   defp do_imports([{module, opts} | rest], types_acc, schema) do
     case ensure_compiled(module) do
       {:module, module} ->
-        [other_def] = module.__absinthe_blueprint__.schema_definitions
+        [other_def] = module.__absinthe_blueprint__().schema_definitions
 
         rejections =
           MapSet.new([:query, :mutation, :subscription] ++ Keyword.get(opts, :except, []))